fix: make ipld macro really work with no_std
Just using `alloc` within the macro doesn't work when the `ipld!` macro is used
outside of this crate. The reason is that `extern crate alloc` would need to be
specified in the consuming crate. We cannot guarantee that. It could result in
errors like:

    error[E0433]: failed to resolve: use of undeclared crate or module `alloc`

A workaround based on [1] is to bring the types from `alloc` that we need into
the scope of the crate and then refer to those from within the macro.

Move the macro tests from unit tests to integration tests to make sure such
problems won't arise again, as now the `ipld!` macro is used from an external
crate.

[1]: https://stackoverflow.com/questions/71675411/refer-to-an-extern-crate-in-macro-expansion/71675639#71675639
